Implemented armor system and fixed health calculation. Corrected weapon logic and changed the character class, having them hold items they pick up rather than just updating the name as it could be useful in later updates

+++ src/Armor.java
@@ -1,5 +1,5 @@
 public class Armor extends Item{
-    public Armor(String name, String type, String description, int attackVal, int speedVal, int defenseVal) {
+    public Armor(String name, String type, String description, int attackVal, int speedVal, double defenseVal) {
         super(name, type, description, 0, attackVal, speedVal, defenseVal);
     }
 }
